NMF moved under Section III, intro refined
The matrix factorization section, which was tacked on at the end, was moved to its rightful home under Section III: Imputation strategies. I combed the Intro more. I added scaling to the tscv_loop formula to see if it improves convergence problems for interpolated and imputed data. So far, not useful.