EDA Reconfigure
Yearly average numbers computed and visualized, with continent hue. Outliers detected, to be removed. Cleaning steps to be moved prior to CLR Transform in predictive analysis. EDA to be migrated out of causal analysis. Once cleaning is completed and visual EDA migrated, we will be ready to conduct DiD analysis in causal notebook.