New Workflow
All exploratory visualizations moved to CoResidence Analysis. This obviates the need for the workflow coming before CLR transformations in CoRes Causal. That is, ll pre-join visualizations and subsequent manual joining of additional demographic and target data to be removed from Casual. New artifact branch point to be assigned in CoResidence Analysis, prior to imputation.