Updated matrix condition source matrix in tscv_loop()
I previously wanted the stronger matrix condition of the XTX (gram matrix), because that is the matrix that gets inverted to solve normal equations. However, upon further research, it appears the canonical approach is to use the scaled X matrix for matrix condition checks, rather than the square of X, as the numbers are more interpretable.