testing after refactoring tscv_loop for potential leakage in means-filling
Got the exploratory model working. This model is not one of the decompose models, so it does not test the new functionality, however, it does test the selection of numeric dtypes within the loop rather than outside of it.

Next steps: eliminate .select_dtypes() preprocessing from outside the loop for all other calls of tscv_loop